menu_money: cast selections compares to int to drop extsh (87.5%)

+++ src/menu_money.cpp
@@ -258,7 +258,7 @@ int CMenuPcs::MoneyCtrlCur()
 		}
 	} else {
 		if ((hold & 8) != 0) {
-			if (this->m_moneyState->selections[mode] != 0) {
+			if ((int)this->m_moneyState->selections[mode] != 0) {
 				this->m_moneyState->selections[mode] = this->m_moneyState->selections[mode] - 1;
 			} else {
 				this->m_moneyState->selections[mode] = 1;
@@ -266,7 +266,7 @@ int CMenuPcs::MoneyCtrlCur()
 			Sound.PlaySe(1, 0x40, 0x7F, 0);
 		} else {
 			if ((hold & 4) != 0) {
-				if (this->m_moneyState->selections[mode] < 1) {
+				if ((int)this->m_moneyState->selections[mode] < 1) {
 					this->m_moneyState->selections[mode] = this->m_moneyState->selections[mode] + 1;
 				} else {
 					this->m_moneyState->selections[mode] = 0;
